chore(sidecar): validate request
